Set the chunk size in amount of frames rather than in microseconds.
With the previous solution (setting chunk size as a time in microseconds),
only a time value corresponding to a round number of frames could be given,
which did not cover all possible values at a given sample rate.
(e.g. at 44100Hz: 10ms is 441 frames, 20ms is 882 frames... but 19 ms for
instance was not allowed as it is not a round number of frames. Also, frames
amounts between 442 and 881 for instance could not be reached as they do not
correspond to a round amount of microseconds).
With chunk size specified as an amount of frames, any value can be given by
the user (provided that the hardware interface supports the given value).